+++ cypress/e2e/ask_user/main.py
@@ -3,7 +3,7 @@ import chainlit as cl
 
 @cl.on_chat_start
 def main():
-    res = cl.AskUser(content="What is your name?", timeout=10).send()
+    res = cl.AskUserMessage(content="What is your name?", timeout=10).send()
     if res:
         cl.Message(
             content=f"Your name is: {res['content']}",